+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `weight` props on `LinkList` component.
+
 ## [22.0.0] - 2018-08-08
 
 Breaking changes: